# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PokeVault API - Pokémon card inventory and NFC management
 * OpenAPI spec version: 0.1.0
 */

export interface DashboardSummary {
  total_cards: number;
  total_invested: number;
  total_market_value: number;
  total_profit_loss: number;
  cards_sold: number;
  total_sold_revenue: number;
  realized_profit_loss: number;
  nfc_tagged_count: number;
}