# lib/api-zod/src/generated/types/nfcUrlCheckResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PokeVault API - Pokémon card inventory and NFC management
 * OpenAPI spec version: 0.1.0
 */

export interface NfcUrlCheckResult {
  url: string;
  byte_length: number;
  max_bytes: number;
  fits: boolean;
  short_url_recommended: boolean;
  /** @nullable */
  message?: string | null;
}