# lib/api-zod/src/generated/types/overlayData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * PokeVault API - Pokémon card inventory and NFC management
 * OpenAPI spec version: 0.1.0
 */
import type { PriceEntry } from "./priceEntry";

export interface OverlayData {
  card_name: string;
  /** @nullable */
  set_name?: string | null;
  /** @nullable */
  card_number?: string | null;
  condition: string;
  /** @nullable */
  market_value: number | null;
  purchase_price: number;
  /** @nullable */
  profit_loss?: number | null;
  /** @nullable */
  image_url?: string | null;
  recent_prices?: PriceEntry[];
}